requirements.txt
requests==2.32.2
geopy==2.4.1
load_dotenv==0.1.0
dependency_injector==4.41.0